move fixtures for creating blueprints into their own directory
